endpoint to get purchases and create purchase

+++ src/main/java/com/alejo/repository/purchases/CardRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CardRepository extends CrudRepository<Card, Long> {
+public interface CardRepository extends CrudRepository<Card, Integer> {
     List<Card> findByUser(User user);
 }
